BUG: ringElem doesn't change size on rotate (iPad)
We've encountered an issue in a project that appears under certain circumstances on ipads. Replication steps:
1. load a page at landscape on an ipad
2. focus to a form field, 
3. switch to portrait 

Result: it doesn't redraw the width and height of the flying focus element. This means that when a field is close to the right hand side of a page, it will create empty space out the side until the field is unfocussed.